chore: complete PR 2E-2H, move semcode format, isolate sm-vm and update tests
