+++ crates/sm-format/Cargo.toml
@@ -5,8 +5,8 @@ edition = "2021"
 publish = false
 
 [dependencies]
-sm-ir = { path = "../sm-ir", default-features = false }
+
 
 [features]
 default = ["std"]
-std = ["sm-ir/std"]
+std = []
