+++ tests/golden_snapshots/public_api/sm_vm_lib.txt
@@ -1,5 +1,5 @@
 source: crates/sm-vm/src/lib.rs
-pub use sm_emit::{
+pub use sm_format::semcode_format::{
 #[cfg(feature = "std")]
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum QuadVal {
